chore(close-button): move out of /common
