+++ packages/components/src/Toast/Toast.tsx
@@ -1,7 +1,7 @@
 import clsx from "clsx";
 import React from "react";
+import CloseButton from "../CloseButton";
 import Text from "../Text";
-import CloseButton from "../common/CloseButton";
 import colorStyles from "../common/Notifications/Notification.module.css";
 import NotificationIcon from "../common/Notifications/NotificationIcon";
 import styles from "./Toast.module.css";
